# core-lending-credit-cards-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-credit-cards
  description: Credit Cards Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/credit-cards
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/cc-revolving-lines/{ccRevolvingLineId}/billing-cycles/{ccBillingCycleId}:
    get:
      tags:
        - CcBillingCycle
      summary: Get a billing cycle by ID
      operationId: getById_4
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcBillingCycleDTO'
    put:
      tags:
        - CcBillingCycle
      summary: Update a billing cycle record
      operationId: update_4
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcBillingCycleDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcBillingCycleDTO'
    delete:
      tags:
        - CcBillingCycle
      summary: Delete a billing cycle record
      operationId: delete_4
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ccBillingCycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/cc-revolving-lines/{ccRevolvingLineId}/billing-cycles:
    get:
      tags:
        - CcBillingCycle
      summary: List or search billing cycles for a revolve line
      operationId: findAll_4
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.ccRevolvingLineId
          in: query
          description: Exact filter for cc revolving line id
          required: false
          schema:
            type: string
        - name: filters.cycleStartDate
          in: query
          description: Exact filter for cycle start date
          required: false
          schema:
            type: string
        - name: filters.cycleEndDate
          in: query
          description: Exact filter for cycle end date
          required: false
          schema:
            type: string
        - name: filters.cycleOpenBalance
          in: query
          description: Exact filter for cycle open balance
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[cycleOpenBalance].from
          in: query
          description: Filter cycle open balance from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[cycleOpenBalance].to
          in: query
          description: Filter cycle open balance to value
          required: false
          schema:
            type: string
        - name: filters.cycleCloseBalance
          in: query
          description: Exact filter for cycle close balance
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[cycleCloseBalance].from
          in: query
          description: Filter cycle close balance from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[cycleCloseBalance].to
          in: query
          description: Filter cycle close balance to value
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCcBillingCycleDTO'
    post:
      tags:
        - CcBillingCycle
      summary: Create a new billing cycle record
      operationId: create_4
      parameters:
        - name: ccRevolvingLineId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CcBillingCycleDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CcBillingCycleDTO'
...
components:
  schemas:
...
    CcBillingCycleDTO:
      required:
        - ccRevolvingLineId
        - cycleCloseBalance
        - cycleEndDate
        - cycleOpenBalance
        - cycleStartDate
      type: object
      properties:
        ccBillingCycleId:
          type: string
          format: uuid
          readOnly: true
        ccRevolvingLineId:
          type: string
          format: uuid
        cycleStartDate:
          type: string
          format: date
        cycleEndDate:
          type: string
          format: date
        cycleOpenBalance:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: -999999999.99
          exclusiveMinimum: false
          type: number
        cycleCloseBalance:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: -999999999.99
          exclusiveMinimum: false
          type: number
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseCcBillingCycleDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CcBillingCycleDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...