# core-lending-credit-cards-web/src/main/java/com/firefly/core/lending/cards/web/controllers/CcBillingCycleController.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.lending.cards.web.controllers;

import org.fireflyframework.core.filters.FilterRequest;
import org.fireflyframework.core.queries.PaginationResponse;
import com.firefly.core.lending.cards.core.services.CcBillingCycleService;
import com.firefly.core.lending.cards.interfaces.dtos.CcBillingCycleDTO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import reactor.core.publisher.Mono;

import java.util.UUID;

@RestController
@RequestMapping("/api/v1/cc-revolving-lines/{ccRevolvingLineId}/billing-cycles")
@Tag(name = "CcBillingCycle", description = "Billing cycle records for a credit card revolve line.")
@RequiredArgsConstructor
public class CcBillingCycleController {

    private final CcBillingCycleService service;

    @GetMapping
    @Operation(summary = "List or search billing cycles for a revolve line")
    public Mono<ResponseEntity<PaginationResponse<CcBillingCycleDTO>>> findAll(
            @PathVariable UUID ccRevolvingLineId,
            @ModelAttribute FilterRequest<CcBillingCycleDTO> filterRequest) {

        return service.findAll(ccRevolvingLineId, filterRequest)
                .map(ResponseEntity::ok);
    }

    @PostMapping
    @Operation(summary = "Create a new billing cycle record")
    public Mono<ResponseEntity<CcBillingCycleDTO>> create(
            @PathVariable UUID ccRevolvingLineId,
            @RequestBody CcBillingCycleDTO dto) {

        return service.create(ccRevolvingLineId, dto)
                .map(ResponseEntity::ok);
    }

    @GetMapping("/{ccBillingCycleId}")
    @Operation(summary = "Get a billing cycle by ID")
    public Mono<ResponseEntity<CcBillingCycleDTO>> getById(
            @PathVariable UUID ccRevolvingLineId,
            @PathVariable UUID ccBillingCycleId) {

        return service.getById(ccRevolvingLineId, ccBillingCycleId)
                .map(ResponseEntity::ok);
    }

    @PutMapping("/{ccBillingCycleId}")
    @Operation(summary = "Update a billing cycle record")
    public Mono<ResponseEntity<CcBillingCycleDTO>> update(
            @PathVariable UUID ccRevolvingLineId,
            @PathVariable UUID ccBillingCycleId,
            @RequestBody CcBillingCycleDTO dto) {

        return service.update(ccRevolvingLineId, ccBillingCycleId, dto)
                .map(ResponseEntity::ok);
    }

    @DeleteMapping("/{ccBillingCycleId}")
    @Operation(summary = "Delete a billing cycle record")
    public Mono<ResponseEntity<Void>> delete(
            @PathVariable UUID ccRevolvingLineId,
            @PathVariable UUID ccBillingCycleId) {

        return service.delete(ccRevolvingLineId, ccBillingCycleId)
                .thenReturn(ResponseEntity.noContent().build());
    }
}